requirements.txt
streamlit
tensorflow
Pillow
opencv-python
numpy